refactor. #63-마이페이지 재렌더링 방지 위해 중첩 레이아웃 적용
