+++ src/pages/[username]/editprofile.tsx
@@ -1,11 +1,17 @@
-import React from "react";
-import MypageLayout from "@/components/Mypage/MypageLayout";
+import React, { ReactElement } from "react";
+import MypageLayout from "@/components/Layout/MypageLayout";
+import { NextPageWithLayout } from "../_app";
+import Layout from "@/components/Layout/Layout";
 
-const EditProfile: React.FC = () => {
+const EditProfile: NextPageWithLayout = () => {
+  return <p>개인정보 변경 ^^</p>;
+};
+
+EditProfile.getLayout = function getLayout(page: ReactElement) {
   return (
-    <MypageLayout>
-      <p>개인정보 변경 ^^</p>
-    </MypageLayout>
+    <Layout>
+      <MypageLayout>{page}</MypageLayout>
+    </Layout>
   );
 };
